requirements.txt
asgiref==3.7.2
Django==4.0.1
django-filter==24.1
djangorestframework==3.14.0
dnspython==2.6.1
drf-yasg==1.21.7
factory-boy==3.3.0
Faker==24.2.0
inflection==0.5.1
Markdown==3.6
packaging==24.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
uritemplate==4.1.1
